Move action mode callback out of the fragment
